Fix HTTPS download of Seattle installer zip
We previously overrode TLS certificate checks for HTTPS downloads
(why did we do that again?), and only verified that the FQDN from
which we attempted to download was on a whitelist. This was a bad
idea, as a MITM could tamper with our DNS and redirect the download
to a site they control.

Also, bump app version to 1.2.